+++ resources/views/etudiants.blade.php
@@ -48,7 +48,7 @@
 													<a href="/etudiants/{{$etudiant->id}}/edit" class="btn btn-primary btn-sm">
 														<i class="bi bi-pencil-square"></i>
 													</a>
-													<a href="/etudiants/{{$etudiant->id}}/delete" class="btn btn-danger btn-sm">
+													<a href="/etudiants/{{$etudiant->id}}/delete" class="btn btn-danger btn-sm confirmation">
 														<i class="bi bi-trash"></i>
 													</a>
 												</td>
